Restructured code and made project a maven project. Added some skeleton test classes.
