+++ README.md
@@ -1,2 +1,2 @@
-# League
+# com.bento909.model.League
 A quick and dirty Java project which reads a CSV file containing sports data and generates a league table.
