+++ src/main/java/com/bento909/Main.java
@@ -1,4 +1,4 @@
-package com.ben;
+package com.bento909;
 
 /**
  * Created by bento on 07/10/16.
