+++ src/main/java/com/bento909/model/League.java
@@ -1,4 +1,5 @@
-package com.ben.model;
+package com.bento909.model;
+
 import java.util.TreeSet;
 
 public class League {
