+++ src/test/java/com/bento909/CsvParserTest.java
@@ -1,0 +1,14 @@
+package com.bento909;
+
+import org.junit.Test;
+
+public class CsvParserTest {
+
+    @Test
+    public void testParseFile() {
+        // Put a test in here which reads a CSV file and tests it's output is what is expected
+        // Fail test below.
+        assert(true == false);
+    }
+
+}
