+++ src/test/java/com/bento909/FileBasedLeagueGeneratorTest.java
@@ -1,0 +1,13 @@
+package com.bento909;
+
+import org.junit.Test;
+
+public class FileBasedLeagueGeneratorTest {
+
+    @Test
+    public void testGenerateLeague() {
+        // Put a test in here which reads a CSV file and tests that the League which is built is as expected
+        // Fail test below.
+        assert(true == false);
+    }
+}
